Bug 2052703.  a=pascalc
Differential Revision: https://phabricator.services.mozilla.com/D310598

+++ image/decoders/nsJXLDecoder.cpp
@@ -418,7 +418,7 @@ nsresult nsJXLDecoder::AllocateFrameBuffers() {
   }
 
   if (mPixelFormat.value() == PixelFormat::Cmyk8 &&
-      !mKBuffer.resize(size.width * size.height)) {
+      !mKBuffer.resize(size_t(size.width) * size.height)) {
     return NS_ERROR_FAILURE;
   }
